Finished the signUp/ login  page
