--- client/src/components/utils/queries.js
@@ -1,3 +1,0 @@
-import { gql } from '@apollo/client';
-
-export const QUERY_LOGIN
