+++ client/src/index.css
@@ -48,13 +48,23 @@ h1 {
   display: flex;
   flex-direction: column;
   justify-content: center;
-  margin-left: 65px;
+  margin-left: 30px;
 }
 .form{
   margin-top: 200px;
   border: 5px solid black;
   width: 300px;
   height: 200px;
+}
+.homeHeader{
+  font-size: 45px;
+}
+.signup-login{
+  display: flex;
+  margin-top: 100px
+}
+.container-box{
+  margin-right: 25px;
 }
