+++ client/src/pages/Home.jsx
@@ -6,12 +6,13 @@ import Main from '../components/Main/index';
 import {QUERY_POST} from '../utils/queries'
 
 function Home(){
-    const element = <h1> What's new in your Hood??</h1>;
+    const element = <h1 className= 'homeHeader'> What's new in your Hood??</h1>;
     const { loading, data } = useQuery(QUERY_POST)
     const postData = data?.getAllPosts || []
     if(loading)return <div> Loading </div>
     return (
         <div>
+            <h1>{element}</h1>
         <Main postData = {postData}/>
         </div>
     )
@@ -19,8 +20,3 @@ function Home(){
 
 export default Home 
 
-// header w/ houses (includes login button)
-// clickable mailbox with neighboorhood events
-// comment feed from main
-// sidebar
-// footer
